코드 수정

+++ src/main/java/com/culture/BAEUNDAY/domain/post/DTO/PostResponse.java
@@ -114,7 +114,7 @@ public class PostResponse {
                 Long userId,
                 String name,
                 String profileImgURL,
-                Double manner
+                Integer manner
         ){
             public UserDTO(User user){
                 this(
